Restructure module and base.
